% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrRank.R
\name{corrRank}
\alias{corrRank}
\title{correlation ranking}
\usage{
corrRank(predictors, outcome)
}
\arguments{
\item{predictors}{a list of predictors}

\item{outcome}{the outcome}
}
\value{
ranking
}
\description{
a function that ranks a list of continuous predictor variables by their correlation with a continuous outcome.
}
\examples{
corrRank(list("BMI", "Blood pressure", "Age"), Cholesterol)
}
