% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot}
\alias{cplot}
\title{correlation plot}
\usage{
cplot(predictors, outcome)
}
\arguments{
\item{predictors}{list of numeric predictor(s), vectors}

\item{outcome}{numeric outcome variable, vector}
}
\value{
list of ggplot2 output(s). List can be indexed to output specific plot(s)
}
\description{
a function that prints bivariate plots of outcome against exposure (from a list of predictors in order)
}
\examples{
exposures <- list(cities$life_expectancy, cities$happiness_level)
plot_list <- cplot(exposures, cities$obesity_level)
plot_list$`Exposure 1`
plot_list$`Exposure 2`
}
