% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrRank.R
\name{corrRank}
\alias{corrRank}
\title{correlation ranking}
\usage{
corrRank(predictors, outcome)
}
\arguments{
\item{predictors}{a list of named predictors}

\item{outcome}{the outcome variable}
}
\value{
a ranking of linear associations by strength, and additional information
}
\description{
a function that ranks a list of continuous predictor variables by their correlation with a continuous outcome.
}
\examples{
corrRank(list(Work = cities$avg_hours_worked_annual,
Pollution = cities$pollution_index, Sunshine = cities$sunshine_hours),
cities$happiness_level)

corrRank(list(Work = cities$avg_hours_worked_annual),
cities$sunshine_hours)
}
