% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot}
\alias{cplot}
\title{correlation plot}
\usage{
cplot(predictors, outcome)
}
\arguments{
\item{predictors}{list of predictors, vectors}

\item{outcome}{outcome variable, vector}
}
\description{
a function that prints bivariate plots of outcome against exposure (from a list of predictors in order)
}
\examples{
exposures <- list(cities$life_expectancy, cities$happiness_level)
cplot(exposures, cities$cities$obesity)
}
